requirements.txt
Flask
gunicorn
google-generativeai>=0.5.0